Refactor, and dequeue second checks when banning or unbanning users
