+++ devvit.yaml
@@ -1,2 +1,2 @@
 name: hive-protect
-version: 1.10.8
+version: 1.10.10
