+++ src/actions/report.ts
@@ -1,5 +1,5 @@
 import { Comment, Post, SettingsValues, TriggerContext } from "@devvit/public-api";
-import { APPROVALS_KEY } from "../hiveProtect.js";
+import { APPROVALS_KEY } from "../handleContentCreation.js";
 import { ProblematicSubsResult } from "../getProblematicItems.js";
 import { replaceAll } from "../utility.js";
 import { addDays } from "date-fns";
