+++ src/main.ts
@@ -1,10 +1,11 @@
 import { Devvit } from "@devvit/public-api";
-import { handleCommentSubmitEvent, handleModActionEvent, handlePostSubmitEvent } from "./hiveProtect.js";
+import { handleCommentSubmitEvent, handlePostSubmitEvent } from "./handleContentCreation.js";
 import { appSettings } from "./settings.js";
 import { handleAppInstallOrUpgradeEvent } from "./installEventHandlers.js";
 import { cleanupDeletedAccounts } from "./cleanupTasks.js";
 import { CLEANUP_JOB, SECOND_CHECK_JOB } from "./constants.js";
 import { handleSecondCheckJob } from "./secondChecks.js";
+import { handleModActionEvent } from "./handleModActions.js";
 
 Devvit.addSettings(appSettings);
